src: move events types into their own module
